Move all AlgorithmHparams classes to shared file (#452)
Move each algorithm's AlgorithmHparams from the algorithm's main implementation file to a shared `composer/algorithms/hparams.py` file

* Clean up imports based on above change. Affected files, besides algorithm implementation files and __init__.py files:

    * `tests/test_load.py` and `tests/test_state.py` needed one modified algorithm import each

    * `tests/algorithms/test_{cutmix,ghost_batchnorm,label_smoothing,selective_backprop}` needed their associated `AlgorithmHparams`  imported from `composer.algorithms` instead of their old locations

* A few method cards (mostly mine) listed the associated `AlgorithmHparams` in the code section. I just removed these mentions for consistency rather than updating the paths
