+++ composer/algorithms/algorithm_registry.py
@@ -3,28 +3,13 @@
 from typing import Dict, List, Type
 
 from composer.algorithms.algorithm_hparams import AlgorithmHparams
-from composer.algorithms.alibi import AlibiHparams
-from composer.algorithms.augmix import AugMixHparams
-from composer.algorithms.blurpool import BlurPoolHparams
-from composer.algorithms.channels_last import ChannelsLastHparams
-from composer.algorithms.colout import ColOutHparams
-from composer.algorithms.cutmix import CutMixHparams
-from composer.algorithms.cutout import CutOutHparams
-from composer.algorithms.factorize import FactorizeHparams
-from composer.algorithms.ghost_batchnorm import GhostBatchNormHparams
-from composer.algorithms.label_smoothing import LabelSmoothingHparams
-from composer.algorithms.layer_freezing import LayerFreezingHparams
-from composer.algorithms.mixup import MixUpHparams
-from composer.algorithms.no_op_model import NoOpModelHparams
-from composer.algorithms.progressive_resizing import ProgressiveResizingHparams
-from composer.algorithms.randaugment import RandAugmentHparams
-from composer.algorithms.sam import SAMHparams
-from composer.algorithms.scale_schedule import ScaleScheduleHparams
-from composer.algorithms.selective_backprop import SelectiveBackpropHparams
-from composer.algorithms.seq_length_warmup import SeqLengthWarmupHparams
-from composer.algorithms.squeeze_excite import SqueezeExciteHparams
-from composer.algorithms.stochastic_depth import StochasticDepthHparams
-from composer.algorithms.swa.hparams import SWAHparams
+from composer.algorithms.hparams import (AlibiHparams, AugMixHparams, BlurPoolHparams, ChannelsLastHparams,
+                                         ColOutHparams, CutMixHparams, CutOutHparams, FactorizeHparams,
+                                         GhostBatchNormHparams, LabelSmoothingHparams, LayerFreezingHparams,
+                                         MixUpHparams, NoOpModelHparams, ProgressiveResizingHparams, RandAugmentHparams,
+                                         SAMHparams, ScaleScheduleHparams, SelectiveBackpropHparams,
+                                         SeqLengthWarmupHparams, SqueezeExciteHparams, StochasticDepthHparams,
+                                         SWAHparams)
 from composer.core.algorithm import Algorithm
 
 registry: Dict[str, Type[AlgorithmHparams]] = {
