+++ composer/algorithms/alibi/__init__.py
@@ -2,7 +2,6 @@
 
 from composer.algorithms.alibi import gpt2_alibi as gpt2_alibi
 from composer.algorithms.alibi.alibi import Alibi as Alibi
-from composer.algorithms.alibi.alibi import AlibiHparams as AlibiHparams
 from composer.algorithms.alibi.alibi import apply_alibi as apply_alibi
 
 _name = 'Alibi'
