+++ composer/algorithms/blurpool/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.blurpool.blurpool import BlurPool as BlurPool
-from composer.algorithms.blurpool.blurpool import BlurPoolHparams as BlurPoolHparams
 from composer.algorithms.blurpool.blurpool import apply_blurpool as apply_blurpool
 from composer.algorithms.blurpool.blurpool_layers import BlurConv2d as BlurConv2d
 from composer.algorithms.blurpool.blurpool_layers import BlurMaxPool2d as BlurMaxPool2d
