+++ composer/algorithms/channels_last/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.channels_last.channels_last import ChannelsLast as ChannelsLast
-from composer.algorithms.channels_last.channels_last import ChannelsLastHparams as ChannelsLastHparams
 from composer.algorithms.channels_last.channels_last import apply_channels_last as apply_channels_last
 
 _name = 'Channels Last'
