+++ composer/algorithms/cutout/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.cutout.cutout import CutOut as CutOut
-from composer.algorithms.cutout.cutout import CutOutHparams as CutOutHparams
 from composer.algorithms.cutout.cutout import cutout_batch as cutout_batch
 
 _name = 'CutOut'
