+++ composer/algorithms/label_smoothing/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.label_smoothing.label_smoothing import LabelSmoothing as LabelSmoothing
-from composer.algorithms.label_smoothing.label_smoothing import LabelSmoothingHparams as LabelSmoothingHparams
 from composer.algorithms.label_smoothing.label_smoothing import smooth_labels as smooth_labels
 
 _name = 'Label Smoothing'
