+++ composer/algorithms/layer_freezing/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.layer_freezing.layer_freezing import LayerFreezing as LayerFreezing
-from composer.algorithms.layer_freezing.layer_freezing import LayerFreezingHparams as LayerFreezingHparams
 from composer.algorithms.layer_freezing.layer_freezing import freeze_layers as freeze_layers
 
 _name = 'Layer Freezing'
