+++ composer/algorithms/mixup/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.mixup.mixup import MixUp as MixUp
-from composer.algorithms.mixup.mixup import MixUpHparams as MixUpHparams
 from composer.algorithms.mixup.mixup import mixup_batch as mixup_batch
 
 _name = 'MixUp'
