+++ composer/algorithms/no_op_model/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.no_op_model.no_op_model import NoOpModel as NoOpModel
-from composer.algorithms.no_op_model.no_op_model import NoOpModelHparams as NoOpModelHparams
 
 _name = 'No-Op Model'
 _class_name = 'NoOpModel'
