+++ composer/algorithms/progressive_resizing/__init__.py
@@ -1,8 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.progressive_resizing.progressive_resizing import ProgressiveResizing as ProgressiveResizing
-from composer.algorithms.progressive_resizing.progressive_resizing import \
-    ProgressiveResizingHparams as ProgressiveResizingHparams
 from composer.algorithms.progressive_resizing.progressive_resizing import resize_batch as resize_batch
 
 _name = 'Progressive Resizing'
