+++ composer/algorithms/scale_schedule/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.scale_schedule.scale_schedule import ScaleSchedule as ScaleSchedule
-from composer.algorithms.scale_schedule.scale_schedule import ScaleScheduleHparams as ScaleScheduleHparams
 
 _name = 'Scale Schedule'
 _class_name = 'ScaleScheduler'
