+++ composer/algorithms/selective_backprop/__init__.py
@@ -1,8 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.selective_backprop.selective_backprop import SelectiveBackprop as SelectiveBackprop
-from composer.algorithms.selective_backprop.selective_backprop import \
-    SelectiveBackpropHparams as SelectiveBackpropHparams
 from composer.algorithms.selective_backprop.selective_backprop import selective_backprop as selective_backprop
 from composer.algorithms.selective_backprop.selective_backprop import \
     should_selective_backprop as should_selective_backprop
