+++ composer/algorithms/seq_length_warmup/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.algorithms.seq_length_warmup.seq_length_warmup import SeqLengthWarmup as SeqLengthWarmup
-from composer.algorithms.seq_length_warmup.seq_length_warmup import SeqLengthWarmupHparams as SeqLengthWarmupHparams
 from composer.algorithms.seq_length_warmup.seq_length_warmup import \
     set_batch_sequence_length as set_batch_sequence_length
