+++ composer/algorithms/swa/__init__.py
@@ -1,6 +1,5 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
-from composer.algorithms.swa.hparams import SWAHparams as SWAHparams
 from composer.algorithms.swa.swa import SWA as SWA
 
 _name = 'SWA'
